refactor(app): extract notification manager
